+++ flutter/lib/src/interface/media_device_info.dart
@@ -1,10 +1,21 @@
 import '../util/move_semantic.dart';
-import 'track_kinds.dart';
+
+/// Kind of a media device.
+enum MediaDeviceKind {
+  /// Audio input device (for example, a microphone).
+  audioinput,
+
+  /// Video input device (for example, a webcam).
+  videoinput,
+
+  /// Audio output device (for example, a pair of headphones).
+  audiooutput,
+}
 
 /// [`MediaDeviceInfo`][1] interface.
 ///
-/// [1]: https://w3.org/TR/mediacapture-streams/#device-info
-abstract class InputDeviceInfo {
+/// [1]: https://w3.org/TR/mediacapture-streams#device-info
+abstract class MediaDeviceInfo {
   /// Returns an unique identifier of the represented device.
   String deviceId();
 
@@ -18,8 +29,8 @@ abstract class InputDeviceInfo {
   ///
   /// This representation of a [`MediaDeviceInfo`][1] is ONLY for input devices.
   ///
-  /// [1]: https://w3.org/TR/mediacapture-streams/#device-info
-  MediaKind kind();
+  /// [1]: https://w3.org/TR/mediacapture-streams#device-info
+  MediaDeviceKind kind();
 
   /// Returns a group identifier of the represented device.
   ///
@@ -28,7 +39,7 @@ abstract class InputDeviceInfo {
   /// representing the speaker and microphone of the same headset have the
   /// same [`groupId`][1].
   ///
-  /// [1]: https://w3.org/TR/mediacapture-streams/#dom-mediadeviceinfo-groupid
+  /// [1]: https://w3.org/TR/mediacapture-streams#dom-mediadeviceinfo-groupid
   String? groupId();
 
   /// Drops the associated Rust struct and nulls the local [Pointer] to it.