+++ flutter/lib/src/interface/media_manager.dart
@@ -1,4 +1,4 @@
-import 'input_device_info.dart';
+import 'media_device_info.dart';
 import 'local_media_track.dart';
 import 'media_stream_settings.dart';
 import '../util/move_semantic.dart';
@@ -10,7 +10,7 @@ import '../util/move_semantic.dart';
 /// tracks for further re-usage.
 ///
 /// [1]: https://w3.org/TR/mediacapture-streams#dom-mediadevices-getusermedia
-/// [2]: https://w3.org/TR/screen-capture/#dom-mediadevices-getdisplaymedia
+/// [2]: https://w3.org/TR/screen-capture#dom-mediadevices-getdisplaymedia
 abstract class MediaManagerHandle {
   /// Obtains [LocalMediaTrack]s objects from local media devices (or screen
   /// capture) basing on the provided [MediaStreamSettings].
@@ -23,7 +23,7 @@ abstract class MediaManagerHandle {
   /// access failed.
   Future<List<LocalMediaTrack>> initLocalTracks(MediaStreamSettings caps);
 
-  /// Returns a list of [InputDeviceInfo] objects representing available media
+  /// Returns a list of [MediaDeviceInfo] objects representing available media
   /// input devices, such as microphones, cameras, and so forth.
   ///
   /// Throws a [StateError] if an underlying object has been disposed, e.g.
@@ -32,7 +32,10 @@ abstract class MediaManagerHandle {
   ///
   /// Throws a [EnumerateDevicesException] if a request of platform media
   /// devices access failed.
-  Future<List<InputDeviceInfo>> enumerateDevices();
+  Future<List<MediaDeviceInfo>> enumerateDevices();
+
+  /// Switches output audio device to the device with the provided [deviceId].
+  Future<void> setOutputAudioId(String deviceId);
 
   /// Drops the associated Rust struct and nulls the local [Pointer] to it.
   @moveSemantics