+++ flutter/lib/src/native/platform/media_device_info.dart
@@ -4,7 +4,7 @@ import 'package:ffi/ffi.dart';
 import 'package:flutter_webrtc/src/model/device.dart';
 import 'package:medea_jason/src/native/ffi/foreign_value.dart';
 
-import 'input_device_info.g.dart' as bridge;
+import 'media_device_info.g.dart' as bridge;
 
 /// Registers functions allowing Rust to operate Dart [MediaDeviceInfo].
 void registerFunctions(DynamicLibrary dl) {