+++ flutter/lib/src/web/media_manager.dart
@@ -1,10 +1,13 @@
-import '../interface/input_device_info.dart';
+import 'package:flutter_webrtc/src/platform/web/video_renderer.dart'
+    as video_renderer;
+
+import '../interface/media_device_info.dart';
 import '../interface/local_media_track.dart';
 import '../interface/media_manager.dart';
 import '../interface/media_stream_settings.dart' as base_settings;
 import '../util/move_semantic.dart';
 import 'exceptions.dart';
-import 'input_device_info.dart';
+import 'media_device_info.dart';
 import 'jason_wasm.dart' as wasm;
 import 'local_media_track.dart';
 import 'media_stream_settings.dart';
@@ -23,14 +26,19 @@ class WebMediaManagerHandle extends MediaManagerHandle {
   }
 
   @override
-  Future<List<InputDeviceInfo>> enumerateDevices() async {
+  Future<List<MediaDeviceInfo>> enumerateDevices() async {
     var tracks = await fallibleFuture(obj.enumerate_devices());
-    return tracks.map((t) => WebInputDeviceInfo(t)).toList();
+    return tracks.map((t) => WebMediaDeviceInfo(t)).toList();
   }
 
   @moveSemantics
   @override
   void free() {
     obj.free();
+  }
+
+  @override
+  Future<void> setOutputAudioId(String deviceId) async {
+    video_renderer.setOutputAudioSinkId(deviceId);
   }
 }