+++ src/platform/dart/media_devices.rs
@@ -5,21 +5,26 @@
 use medea_macro::dart_bridge;
 use tracerr::Traced;
 
-use crate::platform::{
-    dart::utils::{
-        dart_future::FutureFromDart, handle::DartHandle, list::DartList,
+use crate::{
+    api::string_into_c_str,
+    platform::{
+        dart::utils::{
+            dart_future::FutureFromDart, handle::DartHandle, list::DartList,
+        },
+        Error,
     },
-    Error,
 };
 
 use super::{
     constraints::{DisplayMediaStreamConstraints, MediaStreamConstraints},
-    input_device_info::InputDeviceInfo,
+    media_device_info::MediaDeviceInfo,
     media_track::MediaStreamTrack,
 };
 
 #[dart_bridge("flutter/lib/src/native/platform/media_devices.g.dart")]
 mod media_devices {
+    use std::{os::raw::c_char, ptr};
+
     use dart_sys::Dart_Handle;
 
     extern "C" {
@@ -40,6 +45,12 @@ mod media_devices {
         ///
         /// [1]: https://w3.org/TR/mediacapture-streams#mediastreamtrack
         pub fn get_display_media(constraints: Dart_Handle) -> Dart_Handle;
+
+        /// Switches the current output audio device to the device with the
+        /// provided `device_id`.
+        pub fn set_output_audio_id(
+            device_id: ptr::NonNull<c_char>,
+        ) -> Dart_Handle;
     }
 }
 
@@ -54,7 +65,7 @@ mod media_devices {
 ///
 /// [1]: https://tinyurl.com/w3-streams#dom-mediadevices-enumeratedevices
 /// [2]: https://w3.org/TR/mediacapture-streams#mediadevices
-pub async fn enumerate_devices() -> Result<Vec<InputDeviceInfo>, Traced<Error>>
+pub async fn enumerate_devices() -> Result<Vec<MediaDeviceInfo>, Traced<Error>>
 {
     let devices = FutureFromDart::execute::<DartHandle>(unsafe {
         media_devices::enumerate_devices()
@@ -121,4 +132,20 @@ pub async fn get_display_media(
     .map_err(tracerr::wrap!())?;
 
     Ok(DartList::from(tracks).into())
+}
+
+/// Switches the current output audio device to the device with the provided
+/// `device_id`.
+///
+/// # Errors
+///
+/// If output audio device with the provided `device_id` is not available.
+pub async fn set_output_audio_id(
+    device_id: String,
+) -> Result<(), Traced<Error>> {
+    FutureFromDart::execute::<()>(unsafe {
+        media_devices::set_output_audio_id(string_into_c_str(device_id))
+    })
+    .await
+    .map_err(tracerr::wrap!())
 }